Move Stage to being a property of ObjectScrips, not Scrips
The Stage of the scrip is now declared to be part of the association
between the scrip and the queue, and not a property of the scrip itself.
This, in conjunction with the ability to sort scrip execution order on
a per-queue basis, allows for much greater reuse of a single scrip in
multiple contexts.

As 'Disabled' was previously a stage, but should remain a property of
the scrip, the Scrip table gains a Disabled flag.  While this makes it
more consistent with other records' Disabled bits, it does necessitate
additional calls to ->FindAllRows in some call sites to keep the
previous functionality.

    (cherry-picked from rt commit 715367f8c14e16d26604bd1e3bdfc5e38fd413bd)