Remove support for $LinkTransactionsRun1Scrip
The option was initially added to provide backwards compatible behaviour
for 3.4.  We've been recording two link transactions since 3.4.5 and
running scrips on both by default since 3.8.0.  Anyone relying on this
option can adjust their scrips to ignore one of the transactions.

The test file required more adjusting than I would have preferred since
it was written around the assumption of toggling this option.  In the
process of making it pass, I removed the confusing cumulative behaviour
of link_count() by having it reset whenever called.  I also fixed a
minor test bug which called AddLink when it meant to call DeleteLink
(line 170 in the new file).

    (cherry-picked from rt commit 529fa7a722867474d31745710fb6935b8abee54e)