Push all ACL'ing into AddRecord, so ItemsArrayRef respects it
Existing ACL machinery was implemented either on AddRecord (which
applies when data is first pulled from the database), or on Next (which
applies during iteration).  Applying ACLs in Next, however, means that
they are skipped if ->ItemsArrayRef is called.

Move all existing ACL checks to AddRecord, formalizing logic into
->CurrentUserCanSee when relevant.

This change causes test failures for the Dashboards and SavedSearches;
the existing behavior relies on showing you all groups you are a member
of, even if you do not have SeeGroup on those groups.  This behavior
should not change during a stable series; the following commit will add
an ACL exception for their use case.

    (Cherry picked from RT commit f56ca8b54d9f36f73b65f495d1f602a12b7071c0)