Never place the temporary current user in the session
Setting $session{'CurrentUser'} to a different user opens a window
wherein if the request can be aborted, the client will be left with a
session for the other user.  This allows escalation from knowing an RSS
feed link (which is generally just information disclosure) into full
login privileges, which may allow for arbitrary execution of code.  This
vulnerability has been assigned CVE-2015-1464.

    (Cherry picked from RT commit ba3fa035ebf14777dcc33449fba0c2e8a21cd49f)